+++ blocker/CallDirectoryHandler.swift
@@ -51,6 +51,8 @@ class CallDirectoryHandler: CXCallDirectoryProvider {
 
     context.removeAllBlockingEntries()
     context.removeAllIdentificationEntries()
+    
+    print("Successfully reset all numbers list")
   }
 
   private func addNumbersList(
